% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_thresh.R
\name{bayes_thresh}
\alias{bayes_thresh}
\title{Compute the winning threshold using Bayesian Optimization}
\usage{
bayes_thresh(
  mm,
  bounds,
  n_iter,
  init_grid_dt = NULL,
  init_points = 0,
  acq = "ucb",
  kappa = 2.576,
  verbose = FALSE
)
}
\arguments{
\item{mm}{A metric matrix or an object of class 'dist'}

\item{bounds}{A numeric vector of lower and upper bounds. All the sample
points in \code{init_grid_dt} should be in the range of bounds.}

\item{n_iter}{Total number of times the Bayesian Optimization is to repeated.}

\item{init_grid_dt}{A numeric vector of user specified points to sample the
target function.}

\item{init_points}{Number of randomly chosen points to sample the target
function before Bayesian Optimization fitting the Gaussian Process.}

\item{acq}{Acquisition function type to be used. Can be "ucb", "ei" or "poi".
\itemize{
\item \code{ucb} GP Upper Confidence Bound
\item \code{ei} Expected Improvement
\item \code{poi} Probability of Improvement
}}

\item{kappa}{tunable parameter kappa of GP Upper Confidence Bound, to balance exploitation against exploration,
increasing kappa will make the optimized hyperparameters pursuing exploration.}

\item{verbose}{Whether or not to print progress.}
}
\value{
a list of Bayesian Optimization result is returned:
\itemize{
\item \code{Best_Par} a named vector of the best hyperparameter set found
\item \code{Best_Value} the value of metrics achieved by the best hyperparameter set
\item \code{History} a \code{data.table} of the bayesian optimization history
\item \code{Pred} a \code{data.table} with validation/cross-validation prediction for each round of bayesian optimization history
}
}
\description{
This is a simple wrapper for \code{rBayesianOptimization::BayesianOptimization}
that seeks to discover a threshold that produces a small-world, scale-free
network over a given range. Generally 10 or more iterations (including
init points) is needed to find an optimal value. For smaller networks (
under 10000 genes), it should be no problem to run 20 or more iterations
within a few minutes.
}
