% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mmo.R
\name{create_mm}
\alias{create_mm}
\title{Create a metric matrix}
\usage{
create_mm(x, method, ...)
}
\arguments{
\item{x}{A matrix with the -omics data in the rows and the conditions in the columns}

\item{method}{The method passed to dist}

\item{...}{other parameters passed to dist}
}
\description{
For now, a simple wrapper around \code{dist}
}
