% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mmo.R
\name{create_mt}
\alias{create_mt}
\title{Create a metric table}
\usage{
create_mt(x, method, ...)
}
\arguments{
\item{x}{A matrix with the -omics data in the rows and the conditions in the columns}

\item{method}{The method passed to dist}

\item{...}{other parameters passed to dist}
}
\description{
Create a metric table
}
