% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mmo.R
\name{create_omt}
\alias{create_omt}
\title{Create an ordered metric table}
\usage{
create_omt(x, method, sort_ascending = TRUE, ...)
}
\arguments{
\item{x}{A matrix with the -omics data in the rows and the conditions in the columns}

\item{method}{The method passed to dist}

\item{sort_ascending}{Whether the table should be sorted in ascending order}

\item{...}{other parameters passed to dist}
}
\description{
Create an ordered metric table
}
