% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean}
\alias{euclidean}
\title{Compute the Euclidean distance}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\description{
Compute the Euclidean distance
}
