% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ego_graph_from_goi.R
\name{make_ego_graph_from_goi}
\alias{make_ego_graph_from_goi}
\title{Extract the vicinity network for the genes of interest}
\usage{
make_ego_graph_from_goi(graph, order = 1, goi, union_graph = TRUE)
}
\arguments{
\item{graph}{The input igraph graph.}

\item{order}{Integer giving the order of the neighborhood.}

\item{goi}{The gene names for which the calculation is performed.}

\item{union_graph}{Whether to preserve the edges in the original graph.}
}
\description{
Extract the vicinity network for the genes of interest
}
