% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metric.R
\name{Metric}
\alias{Metric}
\title{Create a new metric}
\usage{
Metric(.f, self_distance = .f(1, 1), sort_decreasing, is_symmetric)
}
\arguments{
\item{.f}{A function that takes in two vectors}

\item{self_distance}{The distance between a vector and itself. For the Euclidean
metric, the value is \code{0.0}, and for correlation metrics, the value is \code{1.0}}

\item{sort_decreasing}{Should values be sorted in ascending order (smaller
values implying more similarity) or in descending order (larger values
implying more similarity)}

\item{is_symmetric}{Is the metric \code{f(x, y)} the same as \code{f(y, x)}}
}
\description{
In this package, a metric is used to generally mean the (dis)similarity
between two vectors of genes or proteins.
}
