% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_table.R
\name{metric_table}
\alias{metric_table}
\title{Create a sorted metric table from class \code{dist}}
\usage{
metric_table(x, ...)
}
\arguments{
\item{x}{An object of class \code{dist}}

\item{...}{Extra arguments defined by the specific method}
}
\description{
Create a sorted metric table from class \code{dist}
}
