% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metric_matrix.R
\name{MetricMatrix}
\alias{MetricMatrix}
\title{Create a metric matrix from a matrix of -omics data}
\usage{
MetricMatrix(x, metric, labels = colnames(x))
}
\arguments{
\item{x}{A matrix with the conditions in the rows and the -omics data in the columns}

\item{metric}{An object of class \code{Metric}}

\item{labels}{the labels or names of the columns (-omics)}
}
\value{
An object of class \code{MetricMatrix}
}
\description{
Create a metric matrix from a matrix of -omics data
}
