% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metric_table.R
\name{MetricTable.matrix}
\alias{MetricTable.matrix}
\title{Create a Metric Table from class \code{matrix}}
\usage{
\method{MetricTable}{matrix}(x, metric, sort = FALSE)
}
\arguments{
\item{x}{An object of class \code{matrix}}

\item{metric}{An object of class \code{Metric}}

\item{sort}{should the resulting table be sorted}
}
\description{
Create a Metric Table from class \code{matrix}
}
