% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metric_table.R
\name{MetricTable.MetricTable}
\alias{MetricTable.MetricTable}
\title{Create a Metric Table from class \code{MetricTable}}
\usage{
\method{MetricTable}{MetricTable}(x, sort = FALSE)
}
\arguments{
\item{x}{An object of class \code{MetricTable}}

\item{sort}{should the resulting table be sorted}
}
\description{
Create a Metric Table from class \code{MetricTable}
}
