% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmsym}
\alias{mmsym}
\title{Compute the metric matrix of a symmetric metric}
\usage{
mmsym(X, func)
}
\arguments{
\item{X}{A matrix with the conditions in the rows and the -omics data in the columns}

\item{func}{a metric function that takes in two vectors of equal length}
}
\description{
Compute the metric matrix of a symmetric metric
}
