% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signum_adjacency.R
\name{signum_adjacency}
\alias{signum_adjacency}
\title{Filter a matrix by a simple threshold value}
\usage{
signum_adjacency(x, tau, method = c("less", "greater"))
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{tau}{a cutoff value}

\item{method}{which side of the cutoff value to set to 0}
}
\description{
Filter a matrix by a simple threshold value
}
