% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_utils.R
\name{simple_threshold_seq}
\alias{simple_threshold_seq}
\title{Generate a sequence of threshold values given a 'dist' object}
\usage{
simple_threshold_seq(x, sort_decreasing, length.out = 6)
}
\arguments{
\item{x}{an object of class \code{dist}}

\item{sort_decreasing}{should the resulting table be sorted in descending order}

\item{length.out}{the number of threshold values to return}
}
\description{
Generate a sequence of threshold values given a 'dist' object
}
