% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_loss.R
\name{thresh_loss}
\alias{thresh_loss}
\title{Calculate the loss function for a data frame of thresholds}
\usage{
thresh_loss(thresh_df)
}
\arguments{
\item{thresh_df}{The input threshold data frame}
}
\description{
Calculate the loss function for a data frame of thresholds
}
