% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_stats.R
\name{vertex_stats}
\alias{vertex_stats}
\title{Return basic properties of all vertices in a graph}
\usage{
vertex_stats(graph)
}
\arguments{
\item{graph}{An igraph graph.}
}
\description{
Properties include:
\itemize{
\item vertex number
\item vertex label (if it exists)
\item cluster coefficient
\item degree
\item eccentricity
\item mean distance
\item median distance
}
}
