% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winning_thresh.R
\name{winning_thresh}
\alias{winning_thresh}
\title{Pick the winning threshold the produces a small-world/scale-free network}
\usage{
winning_thresh(
  thresh_df,
  min_r2 = 0.8,
  min_power = 1,
  min_prop_used = 0.4,
  min_prop_big_comp = 0.95
)
}
\description{
Pick the winning threshold the produces a small-world/scale-free network
}
