% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_table.R
\name{metric_table}
\alias{metric_table}
\title{Create a sorted metric table from class \code{dist}}
\usage{
metric_table(x, sort_decreasing)
}
\arguments{
\item{x}{An object of class \code{dist}}

\item{sort_decreasing}{should the resulting table be sorted in descending order}
}
\description{
Create a sorted metric table from class \code{dist}
}
