% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metric_table.R
\name{MetricTable.MetricMatrix}
\alias{MetricTable.MetricMatrix}
\title{Create a Metric Table from class \code{MetricMatrix}}
\usage{
\method{MetricTable}{MetricMatrix}(x, sort = FALSE)
}
\arguments{
\item{x}{An object of class \code{MetricMatrix}}

\item{sort}{should the resulting table be sorted}
}
\description{
Create a Metric Table from class \code{MetricMatrix}
}
