% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_utils.R
\name{simple_threshold_seq}
\alias{simple_threshold_seq}
\title{Generate a sequence of threshold values given a 'dist' object}
\usage{
simple_threshold_seq(x, sort_decreasing, length.out = 6)
}
\arguments{
\item{x}{an object of class \code{dist}}

\item{sort_decreasing}{should the resulting table be sorted in descending order}

\item{length.out}{the number of threshold values to return}
}
\value{
A sequence of threshold values.
}
\description{
A simple rule is used to get the first and last thresholds. If there are \code{m}
variables, then there are \code{m(m-1)/2} pairs of variables in the sorted metric
table, M. Take the first threshold to be \code{M[m-1]} and the last threshold to be
\verb{M[150m]} \insertCite{petereit2016petal}{petal2}.
}
\examples{
\dontrun{
x <- metric_matrix(dat, "spearman")
thresh <- simple_threshold_seq(x, sort_decreasing=TRUE)
}
}
\references{
\insertAllCited{}
}
